Consolidate remaining startup in phase 3
This commit moves the remaining pieces of startup into a "phase 3". This
consists of everything after security manager is initialized. Phase 3
is where the node is constructed and the system becomes multi threaded.
