+++ server/src/main/java/org/elasticsearch/bootstrap/Elasticsearch.java
@@ -10,18 +10,29 @@ package org.elasticsearch.bootstrap;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.apache.logging.log4j.core.LoggerContext;
+import org.apache.logging.log4j.core.config.Configurator;
+import org.apache.lucene.util.Constants;
+import org.apache.lucene.util.StringHelper;
 import org.elasticsearch.ElasticsearchException;
+import org.elasticsearch.Version;
 import org.elasticsearch.cli.ExitCodes;
-import org.elasticsearch.cli.UserException;
+import org.elasticsearch.common.filesystem.FileSystemNatives;
 import org.elasticsearch.common.io.stream.InputStreamStreamInput;
 import org.elasticsearch.common.logging.LogConfigurator;
 import org.elasticsearch.common.network.IfConfig;
 import org.elasticsearch.common.settings.KeyStoreWrapper;
 import org.elasticsearch.common.settings.SecureSettings;
 import org.elasticsearch.common.settings.Settings;
+import org.elasticsearch.common.transport.BoundTransportAddress;
+import org.elasticsearch.core.IOUtils;
 import org.elasticsearch.core.SuppressForbidden;
 import org.elasticsearch.env.Environment;
 import org.elasticsearch.jdk.JarHell;
+import org.elasticsearch.monitor.jvm.HotThreads;
+import org.elasticsearch.monitor.jvm.JvmInfo;
+import org.elasticsearch.monitor.os.OsProbe;
+import org.elasticsearch.monitor.process.ProcessProbe;
 import org.elasticsearch.node.Node;
 import org.elasticsearch.node.NodeValidationException;
 
@@ -32,9 +43,10 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.security.Permission;
 import java.security.Security;
+import java.util.List;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
 
-import static org.elasticsearch.bootstrap.Bootstrap.initializeNatives;
-import static org.elasticsearch.bootstrap.Bootstrap.initializeProbes;
 import static org.elasticsearch.bootstrap.BootstrapInfo.USER_EXCEPTION_MARKER;
 import static org.elasticsearch.bootstrap.BootstrapSettings.SECURITY_FILTER_BAD_DEFAULTS_SETTING;
 
@@ -55,20 +67,9 @@ class Elasticsearch {
 
         try {
             BootstrapState state = initPhase2(serverArgs);
-            Bootstrap.init(serverArgs, state);
-
-            err.println(BootstrapInfo.SERVER_READY_MARKER);
-            if (serverArgs.daemonize()) {
-                out.close();
-                err.close();
-            } else {
-                startCliMonitorThread(System.in);
-            }
-
+            initPhase3(serverArgs, state, err, out);
         } catch (NodeValidationException e) {
             exitWithUserException(err, ExitCodes.CONFIG, e);
-        } catch (UserException e) {
-            exitWithUserException(err, e.exitCode, e);
         } catch (Throwable t) {
             exitWithUnknownException(err, t);
         }
@@ -202,11 +203,7 @@ class Elasticsearch {
         // initialize probes before the security manager is installed
         initializeProbes();
 
-        Runtime.getRuntime().addShutdownHook(new Thread(() -> {
-            if (Bootstrap.INSTANCE != null) {
-                Bootstrap.INSTANCE.shutdown();
-            }
-        }));
+        Runtime.getRuntime().addShutdownHook(new Thread(Elasticsearch::shutdown));
 
         // look for jar hell
         final Logger logger = LogManager.getLogger(JarHell.class);
@@ -223,6 +220,149 @@ class Elasticsearch {
         );
 
         return new BootstrapState(nodeEnv, keystore, spawner);
+    }
+
+    /**
+     * Third phase of initialization.
+     *
+     * <p> Phase 3 consists of everything after security manager is initialized. Up until now, the system has been single
+     * threaded. This phase can spawn threads, write to the log, and is subject ot the security manager policy.
+     *
+     * <p> At the end of phase 3 the system is ready to accept requests and the main thread is ready to terminate. This means:
+     * <ul>
+     *     <li>The node components have been constructed and started</li>
+     *     <li>Cleanup has been done (eg secure settings are closed)</li>
+     *     <li>At least one thread other than the main thread is alive and will stay alive after the main thread terminates</li>
+     *     <li>The parent CLI process has been notified the system is ready</li>
+     * </ul>
+     *
+     * @param args arguments sent to this server
+     * @param state state from initializing security manager
+     * @throws IOException if a problem with filesystem or network occurs
+     * @throws NodeValidationException if the node cannot start due to a node configuration issue
+     */
+    private static void initPhase3(ServerArgs args, BootstrapState state, PrintStream err, PrintStream out) throws IOException, NodeValidationException {
+        checkLucene();
+
+        Node node = new Node(state.environment()) {
+            @Override
+            protected void validateNodeBeforeAcceptingRequests(
+                final BootstrapContext context,
+                final BoundTransportAddress boundTransportAddress,
+                List<BootstrapCheck> checks
+            ) throws NodeValidationException {
+                BootstrapChecks.check(context, boundTransportAddress, checks);
+            }
+        };
+        INSTANCE = new Elasticsearch(state.spawner, node);
+
+        // any secure settings must be read during node construction
+        IOUtils.close(state.secureSettings());
+
+        INSTANCE.start();
+
+        if (args.daemonize()) {
+            LogConfigurator.removeConsoleAppender();
+        }
+
+        // DO NOT MOVE THIS
+        // Signaling readiness to accept requests must remain the last step of initialization. Note that it is extremely
+        // important closing the err stream to the CLI when daemonizing is the last statement since that is the only
+        // way to pass errors to the CLI
+        err.println(BootstrapInfo.SERVER_READY_MARKER);
+        if (args.daemonize()) {
+            out.close();
+            err.close();
+        } else {
+            startCliMonitorThread(System.in);
+        }
+    }
+
+    /**
+     * Initialize native resources.
+     *
+     * @param tmpFile          the temp directory
+     * @param mlockAll         whether or not to lock memory
+     * @param systemCallFilter whether or not to install system call filters
+     * @param ctrlHandler      whether or not to install the ctrl-c handler (applies to Windows only)
+     */
+    static void initializeNatives(final Path tmpFile, final boolean mlockAll, final boolean systemCallFilter, final boolean ctrlHandler) {
+        final Logger logger = LogManager.getLogger(Elasticsearch.class);
+
+        // check if the user is running as root, and bail
+        if (Natives.definitelyRunningAsRoot()) {
+            throw new RuntimeException("can not run elasticsearch as root");
+        }
+
+        if (systemCallFilter) {
+            /*
+             * Try to install system call filters; if they fail to install; a bootstrap check will fail startup in production mode.
+             *
+             * TODO: should we fail hard here if system call filters fail to install, or remain lenient in non-production environments?
+             */
+            Natives.tryInstallSystemCallFilter(tmpFile);
+        }
+
+        // mlockall if requested
+        if (mlockAll) {
+            if (Constants.WINDOWS) {
+                Natives.tryVirtualLock();
+            } else {
+                Natives.tryMlockall();
+            }
+        }
+
+        // listener for windows close event
+        if (ctrlHandler) {
+            Natives.addConsoleCtrlHandler(new ConsoleCtrlHandler() {
+                @Override
+                public boolean handle(int code) {
+                    if (CTRL_CLOSE_EVENT == code) {
+                        logger.info("running graceful exit on windows");
+                        shutdown();
+                        return true;
+                    }
+                    return false;
+                }
+            });
+        }
+
+        // force remainder of JNA to be loaded (if available).
+        try {
+            JNAKernel32Library.getInstance();
+        } catch (Exception ignored) {
+            // we've already logged this.
+        }
+
+        Natives.trySetMaxNumberOfThreads();
+        Natives.trySetMaxSizeVirtualMemory();
+        Natives.trySetMaxFileSize();
+
+        // init lucene random seed. it will use /dev/urandom where available:
+        StringHelper.randomId();
+
+        // init filesystem natives
+        FileSystemNatives.init();
+    }
+
+    static void initializeProbes() {
+        // Force probes to be loaded
+        ProcessProbe.getInstance();
+        OsProbe.getInstance();
+        JvmInfo.jvmInfo();
+        HotThreads.initializeRuntimeMonitoring();
+    }
+
+    static void checkLucene() {
+        if (Version.CURRENT.luceneVersion.equals(org.apache.lucene.util.Version.LATEST) == false) {
+            throw new AssertionError(
+                "Lucene version mismatch this version of Elasticsearch requires lucene version ["
+                    + Version.CURRENT.luceneVersion
+                    + "]  but the current lucene version is ["
+                    + org.apache.lucene.util.Version.LATEST
+                    + "]"
+            );
+        }
     }
 
     /**
@@ -319,5 +459,55 @@ class Elasticsearch {
             builder.setSecureSettings(secureSettings);
         }
         return new Environment(builder.build(), configDir);
+    }
+
+    // -- instance
+
+    private static volatile Elasticsearch INSTANCE;
+
+    private final Spawner spawner;
+    private final Node node;
+    private final CountDownLatch keepAliveLatch = new CountDownLatch(1);
+    private final Thread keepAliveThread;
+
+    private Elasticsearch(Spawner spawner, Node node) {
+        this.spawner = spawner;
+        this.node = node;
+        this.keepAliveThread = new Thread(() -> {
+            try {
+                keepAliveLatch.await();
+            } catch (InterruptedException e) {
+                // bail out
+            }
+        }, "elasticsearch[keepAlive/" + Version.CURRENT + "]");
+    }
+
+    private void start() throws NodeValidationException {
+        node.start();
+        keepAliveThread.start();
+    }
+
+    private static void shutdown() {
+        if (INSTANCE == null) {
+            return; // never got far enough
+        }
+        var es = INSTANCE;
+        try {
+            IOUtils.close(es.node, es.spawner);
+            LoggerContext context = (LoggerContext) LogManager.getContext(false);
+            Configurator.shutdown(context);
+            if (es.node != null && es.node.awaitClose(10, TimeUnit.SECONDS) == false) {
+                throw new IllegalStateException(
+                    "Node didn't stop within 10 seconds. " + "Any outstanding requests or tasks might get killed."
+                );
+            }
+        } catch (IOException ex) {
+            throw new ElasticsearchException("failed to stop node", ex);
+        } catch (InterruptedException e) {
+            LogManager.getLogger(Elasticsearch.class).warn("Thread got interrupted while waiting for the node to shutdown.");
+            Thread.currentThread().interrupt();
+        } finally {
+            es.keepAliveLatch.countDown();
+        }
     }
 }
